rate Limit OTP

+++ src/otp/otp.controller.js
@@ -1,19 +1,38 @@
 import Otp from "./model.js";
-
 import generateOtp from "../util/generateOtp.js";
 import dotenv from "dotenv";
 import { hashData, verifyHash } from "../util/hashData.js";
 import sendEmail from "../util/sendEmailOtp.js";
+import rateLimit from "express-rate-limit";
 
 dotenv.config();
 
 const { EMAIL_USER } = process.env;
+
+const otpTimeStamps = new Map();
+
+export const otpRateLimiter = rateLimit({
+    windowMs: 1 * 60 * 1000, // 1 minute window
+    max: 5, // Limit each IP to 5 requests per `windowMs`
+    message: "Too many OTP requests, please try again later",
+});
 
 export const sendOtp = async (email, subject, message, duration =1) => {
     try {
         if (!email || !subject || !message) {
             throw new Error("Email, subject, and message are required");
         }
+
+
+        const lastOtpTimestamp = otpTimeStamps.get(email);
+        const currentTime = new Date().getTime();
+
+        if(lastOtpTimestamp && currentTime - lastOtpTimestamp < 60000) {
+            throw new Error("please wait for 1 minute before requesting again");
+        }
+
+        otpTimeStamps.set(email, currentTime);
+        
 
         await Otp.deleteOne({ email });
 
@@ -48,9 +67,10 @@ export const sendOtp = async (email, subject, message, duration =1) => {
 }
 
 export const verifyOtp = async (email, otp) => {
-    if (!(otp && email)) {
-        throw new Error("provide email and otp");
-    }
+    try {
+        if (!(otp && email)) {
+            throw new Error("provide email and otp");
+        }
     const otpDocument = await Otp.findOne({ email });
 
     if(!otpDocument) {
@@ -61,13 +81,21 @@ export const verifyOtp = async (email, otp) => {
     if(expiresAt < new Date()) {
         throw new Error("OTP expired");
     }
+
     const hashedOtp = otpDocument.otp;
     const isOtpMatch = await verifyHash(otp, hashedOtp);
-    return isOtpMatch;
 
-    
+    if(!isOtpMatch) {
+        throw new Error("OTP invalid");
+    }
 
-   
+        await Otp.deleteOne({ email });
+        return true;
+
+    } catch (error) {
+        console.error("OTP verification failed:", error);
+        throw new Error("OTP verification failed: " + error.message);
+    }
 }
 
 export const deleteOtp = async (email) => {
